Use an arma::field for the ledger instead of the overengineered Readable/Writable stuff

+++ src/prisoners/Stochastic.cpp
@@ -7,7 +7,7 @@ Stochastic::Stochastic(float coopProb, uint seed)
 
 Action Stochastic::decide(const PrisonerKnowledge& p) const {
 	std::mt19937 gen;
-	std::seed_seq sseq {_seed, p.time, p.thisIndex, p.theirIndex};
+	std::seed_seq sseq {_seed, (uint)p.time, p.thisIndex, p.theirIndex};
 	gen.seed(sseq);
 	std::bernoulli_distribution dist(_coopProb);
